CDH-4071. Introduce Ivy 'test' configuration.
Switch the test dependency in ql on
hadoop-mapreduce-client-jobclient-*-tests.jar to be part of a new
test configuration in Ivy, distinct from the default/compile
configuration. This'll help us keep the test dependencies from
bleeding over into the compile dependencies, and is a starting point
for the overall reworking of the dependency management.

For right now, the only jar in the test configuration that actually
makes it onto the classpath is the one explicitly specified, since
this is the only Hadoop dependency being handled directly via Ivy,
rather than indirectly through the tarball, but that'll change. The
make-pom target in build-common.xml has also been tweaked to support
adding dependencies in the test configuration as test-scoped
dependencies to the generated POMs.

Reason: Improvement
Author: Andrew Bayer
Ref: CDH-4071