CDH-16383: HIVE-3181: getDatabaseMajor/Minor version does not return values (Szehon via Xuefu, reviewed by Navis)
Minor conflicts due to code differences of Hive 0.10 with Hive 0.13 branches.  Fixed a test failure in TestJdbcDriver2, the expected exception type had changed but not updated, as test was marked for skip in CDH4-0.10 build.  Test was correctly updated in cdh5-0.12 branch.  Will update configuration to run the test in this build as well.