CDH-20531: For local jobs, restore framework before restoring job launcher address
When an MR job is run in local mode, it's non-local attributes in
the config (like job launcher addresses and the framework) must be
saved in the context, and restored once the job is done executing.

Since the isMR2() method uses the framework to determine if it's MR2,
we should restore it before we attempt to restore the job launcher address.
Otherwise, we erroneously restore the MR1 property when MR2 property
needs to be restored.