CDH-21086 : Fix union_remove_17 q-test
The union_remove_17.q.out golden file was updated by a backport of HIVE-6861
(which fixes bunch of hadoop-2 related tests).

However, we are seeing it fail in cdh5.2, where ReduceSinkOperator has duplicate column names.
There is a fix upstream (https://issues.apache.org/jira/browse/HIVE-4867) which addresses duplicate column names,
but changes a whole lot of files (both source and test) and is a pretty hairy merge.

Since it's a purely performance fix ("unnecessary overhead for shuffling", acc to jira) and this is the
only related test failure in cdh5.2, I'm inclined to just fix this locally for cdh5.2