CDH-20832 - Operation should not create driver with parent config
In "CDH-20161: Forward port CDH-13524", operation creates driver
with parent session config. In original patch (CDH-13524), this made sense.
However, now we should just use the sqlOperationConf passed to the prepare
method.

For example, in Hue this was manifesting as MR job not getting confOverlay,
since no confOveray is set for parent session:
  use default;
  select count(*) from sample_07; // picks from config from "use default",
                                  // which doesn't have confOverlay set.