CDH-21446 : Fix TestHS2ThreadAllocation.testHS2StabilityOnLargeConnections
We have a downstream-only Thrift patch that introduces a new param (requestTimeout),
that drops the client connection after requestTimeout, instead of waiting indefinitely for a
thread to free up for handoff.

I had removed this param from HS2 Thrift service when resolving conflicts for
HIVE-7353.

This change re-instates using requestTimeout.