CDH-28929 - CDH5.4.x Hive build broken
This issue is caused by backporting HIVE-10437, which also depends on HIVE-9298.
Hive-9298 is a fairly large patch, and has its own dependencies. Since HIVE-10437
just needs one line from the latter, here, instead of backporting HIVE-9298, we
just manually added that one line.
In future, if we are doing backporting or rebasing, this should be resolved
automatically.