CLOUDERA-BUILD: CDH-40690: [test regression] HIVE-13696
FairSchedulerShim update does not check if the username is 'null'. Adds guard to ensure that NPEs don't
occur when this is the case. The upstream patch for HIVE-13696 will also be updated to contain the fix for this issue.

Change-Id: I32af7bd2e3734273b7d1b1546b8b19ba87488816
(cherry picked from commit 2d965000a24e450cb5e55464bdb45f66e2bb9b86)