CDH-42036: Backport HIVE-11487: Add getNumPartitionsByFilter api in metastore api
This backport is not bringing any new Thrift API change, but only the public methods needed
by HIVE-13884 to work.

Also, there was another single file that it is brought on this change, PartFilterExprUtil.java. The
rest of the patch containing it it is for HBASE changes, and we do not support it yet.

Change-Id: Ia577dd456f0404f9a830c62100d38e2850dd9709
(cherry picked from commit 29db983c173293c33db43847439ab2d68942f7e4)