CLOUDERA-BUILD : Fix for test failure inputwherefalse.q
The fix for HIVE-14614 works in upstream but doesn't work in CDH due to differences in the configuration values for upstream tests and CDH as of this change. This fix is needed so that inputwherefalse.q test passes. The backport for the HIVE-14614 was reverted earlier and this standalone fix is needed to fix the test failures

Change-Id: Icadfdccc0bdafca58ce7adfb3907a20c849164a3