CDH-56326: HIVE-17128: Operation Logging leaks file descriptors as the log4j Appender is never closed (Andrew Sherman, reviewed by Aihua Xu and Peter Vary)
Previously HIVE-16061 and HIVE-16400 changed Operation Logging to use the Log4j2
RoutingAppender to automatically output the log for each query into an
individual operation log file.  As log4j does not know when a query is finished
it keeps the OutputStream in the subordinate Appender open even when the query
completes.  The stream holds a file descriptor and so we leak file descriptors.
Note that we are already careful to close any streams reading from the operation
log file.  To fix this we use a technique described in the comments of
LOG4J2-510 which uses reflection to close the subordinate Appender.  We use this
to close the per-query subordinate Appenders from both LogDivertAppender and
LogDivertAppenderForTest.  The test in TestOperationLoggingLayout is extended to
check that the Appenders are closed when a query completes.

This change also pulls in a one line test fix from HIVE-14612:
TestOperationLoggingLayout.testSwitchLogLayout failure (Hari Subramaniyan, reviewed by Prasanth Jayachandran)
to make the merge cleaner

Change-Id: I30a91f3ff7485b50817afcbf85612a72d71c97fa