CDH-57253 : HIVE-17368 : DBTokenStore fails to connect in Kerberos enabled remote HMS environment (Vihang Karajgaonakar, reviewed by Aihua Xu and Janaki Lahorani)
Notes:
In CDH5 branch there were conflicts while backporting the change. Specifically, CDH5 does not have
HiveDelegationTokenManager.java but instead the code to get the delegation tokens is implemented
in HadoopThriftAuthBridge.java. The upstream fix in HiveDelegationTokenManager.java was manually
ported to HadoopThriftAuthBridge.java to resolve this conflict. Also, SessionState.java does
not have a method called uncacheDatanucleusClassLoaders, so that part of the upstream patch
was not needed.

Change-Id: I78ad1e6282a34f8f785355d18be8805a96e1882f