CDH-58519: HIVE-17405: HoS DPP ConstantPropagate should use ConstantPropagateOption.SHORTCUT (Sahil Takiar, reviewed by Rui Li)
The method ConstantPropagateProcCtx#getPropagatedConstants was modified to avoid
an ArrayIndexOutOfBoundsException that was being thrown when processing union
operators. The modification was to simply check if the array is empty or not.
This is a Cloudera specific change, the upstream implementation of this method
is very different from the current CDH version, which is why this simple fix
was added instead of backporting other upstream patches, see CDH-58519 for
details.

(cherry picked from commit e3c71b009bdeb2191f12d60f288486a5d628ecae)

Conflicts:
	ql/src/test/results/clientpositive/spark/skewjoinopt10.q.out
	ql/src/test/results/clientpositive/spark/skewjoinopt12.q.out
	ql/src/test/results/clientpositive/spark/skewjoinopt14.q.out
	ql/src/test/results/clientpositive/spark/skewjoinopt16.q.out
	ql/src/test/results/clientpositive/spark/skewjoinopt17.q.out
	ql/src/test/results/clientpositive/spark/skewjoinopt2.q.out
	ql/src/test/results/clientpositive/spark/spark_dynamic_partition_pruning.q.out
	ql/src/test/results/clientpositive/spark/spark_vectorized_dynamic_partition_pruning.q.out

Change-Id: I443d3b2c56d2620a54e9836be449dd0615c17cf2