CDH-51004: HIVE-16395: ConcurrentModificationException on config object in HoS.
In SPARK-2546 a config called spark.hadoop.cloneConf was added that is
set to false by default. When set to true it clones a new JobConf object
for each Spark Task, which avoids any thread safety issues.
In Spark it's set to false by default for performance considerations.
On the Hive side we now think the cost of cloning a Spark configuration
is a price worth paying to avoid concurrency problems.
We therefore force spark.hadoop.cloneConf to "true" in the Spark config
unless the user has set this explicitly in the Hive config.
Add a new test that forces a Spark config to be generated from a Hive config
and check that the spark.hadoop.cloneConf config value is as expected.

Change-Id: I452374c9bdb6cdbab0367668168b4820bc20ad50