CDH-58655: HIVE-15543: Don't try to get memory/cores to decide parallelism when Spark dynamic allocation is enabled (Reviewed by Rui) CDH-61154: HIVE-15796: HoS: poor reducer parallelism when operator stats are not accurate (Chao Sun, reviewed by Xuefu Zhang) CDH-59904: HIVE-17010: Fix the overflow problem of Long type in SetSparkReducerParallelism (Liyun Zhang reviewed by Li Rui, Sun Chao, Ferdinand Xu) CDH-58617: HIVE-17114: HoS: Possible skew in shuffling when data is not really skewed (Rui reviewed by Chao) CDH-60781: HIVE-15526: Some tests need SORT_QUERY_RESULTS (Rui reviewed by Xuefu) CDH-61155: HIVE-17868: Add 'order by' to spark_local_queries.q to make tests deterministic
This is a bunch of HoS backports that were found as part of CDH-42541
(investigate and document Hive on Spark Support)
They are all combined into one as they affect similar files.
In addition 2 upstream fixes are needed for test stability.
The upstream code fixes merge cleanly by cherry-pick.
The .q.out files needed to be regenerated.

Change-Id: Ifdb6da19baeb693d468e5bf949b516ab324a1c10