CDH-58943: HIVE-17789: Flaky test: TestSessionManagerMetrics.testAbandonedSessionMetrics has timing related problems
The test is waiting for a worker thread to be timed out. The time after which
the timeout should happen in 3000 ms. The test waits for 3200 ms, and
sometimes this is not enough as the timeout happens asynchronously.  Use a
counter-limited loop to wait for the counter to change; this is quicker than
just sleeping for a long time.

Change-Id: Iaa5809108676d85e91a1b068c7efe71816154d26