HIVE-17870: Update NoDeleteRollingFileAppender to use Log4j2 api Fix is part of NAV-3691
In log4j2 functionality that was previously the responsibility of the
Appender is now split into multiple interfaces (RolloverStrategy,
AbstratcMananager, Appender).  To replace NoDeleteRollingFileAppender we
provide a RolloverStrategy called SlidingFilenameRolloverStrategy which does
the following: (1) creates log file names by appending a millisecond
timestamp to a template (2) when log file rollover occurs the old files are
not renamed, instead logging switches to a file with a new timestamp.

Change log4j2.version to be 2.8.2 as LOG4J2-907 was needed to make the change
work (The log4j2 version is is now the same as standalone-metastore, this
is not a requirement, but nice to have).
Delete the obsolete log4j based NoDeleteRollingFileAppender

Change-Id: I99535850f703547c41d249a32572b821aaf24354