CLOUDERA-BUILD: CDH-62576: Re-writes for TestSparkPerfCliDriver query12,20,42,84,85,91,96,98
Simple re-writes for the above queries so that all order by
columns are always selected.

Ordering by columns not in the select clause is not
supported in CDH Hive. It is supported upstream (see
HIVE-15160). We could backport that JIRA, but its CBO
specific anyway.

Instead, we make a few slight modifications to the queries
so that they compile. There are NO functional changes to
these queries, they still do the exact same thing; the only
change is what column references are used.

Change-Id: Ie501d7eaeacdc5a0d60c642aeff98a250b7c7022