CLOUDERA-BUILD: CDH-63698: Hack to fix NoSuchMethodErrors in TestSparkClient
The JIRA has more details as to why this is necessary. In
short, there is a CDH wide initiative to upgrade the
version of fasterxml.jackson. This was done for Spark
recently, but caused the TestSparkClient tests to start
failing with a NoSuchMethodError due to jar version
conflicts.

The fix in this patch just adds a test dependency to
spark-client for the upgraded version of
fasterxml.jackson. We can revert this change once we have
completed the upgrade to the new version of jackson for
Hive.

Change-Id: I4abd2cfe14967cb7a8320d75a9a60ef0d573be65