CLOUDERA-BUILD: CDH-66444: TestJdbcWithMiniKdcCookie.testCookieNegative is consistently failing
Corresponding JIRA has the details, basically HADOOP-9747
changed the message in the exception that gets thrown when
a user tries to run login via a UGI, but the user doesn't
exist. This test parses the exception message so it expects
it contain a certain string, this commit just changes that
string.

Change-Id: I5beac1990984477a3c4adf3e3cd8d9deb2278006