CDH-68045 : HIVE-19384: Vectorization: IfExprTimestamp* do not handle NULLs correctly (Matt McCline, reviewed by Teddy Choi)
This commit also includes the upstream addendum for this patch. Additionally, TestVectorIfStatement was modified
so that it can be run in CDH code base. The VectorExtractRow class in CDH extracts a writable object from a batch
given a row number and column number. If the caller uses this object directly instead of creating a copy of it
the reference can store a different value a later point of time. Fortunally, in a real query execution VectorExtractRow
is used in VectorFileSinkOperator and it processes the extracted row before moving to the next row. In order to make
the test work without backporting more large patches and their dependencies I had to write the copying logic in the test

Also had to modify Test util class VectorRandomRowSource to exclude the DataTypePhysicalVariation related change which is
not available in CDH.

Change-Id: I414992b4e4b9841fd12a0eb0cea8be519780fbfb
(cherry picked from commit efeea739568f313b0a9c881b777abb6b6bce08c9)