CLOUDERA-BUILD: CDH-67053: Revert jackson version upgrade for testutils/ptest2/pom.xml
==C5_APPROVED_BUGFIX==

CDH-67053 upgraded the jackson version for ptest, which
caused our test infra to break. The reason is that the
ptest2 codebase is not connected to the root pom.xml
so jackson-mapper-asl.1.9.version is not defined for
ptest2.

Ptest2 is not product code nor is it usuable by
customers. It tests Hive tests in parallel across a cluster
of nodes.

Change-Id: I13ac2c9c477af50299d90f1560577ff478a6f8a9