CDH-74733 : HIVE-20307 : Add support for filterspec to the getPartitions with projection API (Vihang Karajgaonkar, reviewed by Andrew Sherman)
The change had some conflicts and was not a clean backport. Following 
changes were needed to make it work in the CDH.

1. CDH does not have catname, so all the methods where catalog name name
is passed around in upstream version of this patch were modified to remove
that argument

2. Upstream had a util method called MetaStoreUtils.makePartNameMatcher 
which was reused by the original patch. The method was not available in
CDH and hence needed to add the method in this patch.

Change-Id: I2f8a6f4016df3f8bdcd05638097476d38fb851aa