CDH-77745: Attempt to move table between encryption zones corrupts metadata
The root cause of this problem is that the HDFS move happens after the
transaction is complete, so when the HDFS move fails, there is code
to manually rollback the transaction that was already committed. The fix
is to call alterTableUpdateTableColumnStats, which also calls alterTable.
This will leave the data in the TAB_COL_STATS correct which will allow
deletion of the table.

Tested manually on a secured cluster both with the reproduce scenario
and with a partitioned table. Hhowever the data in PART_COL_STATS does
not remain as this code path does not call
updatePartitionColumnStatistics but does not prevent deletion of the
table.

Change-Id: I13061baaeb770135ffba20835d00a9bcd58cabe6