CDH-78695: Fix serialization issue in MapReduce
A new timer to calculate the stats for all the API calls was added
to PerfLogger (HIVE-21044 and HIVE-21045). However, when backporting
this change to CDH5, the Metrics class is implemented differently.
This caused the PerfLogger class to be not-serializeable due to missing
libraries. This change fixes this for CDH5 by implementing the change
in the old-CDH5-style Metrics.

Change-Id: If3549abcf0dd5b297268cbd400c4348f01cfcd9d